desktop-app/requirements.txt
# GUI Framework
customtkinter==5.2.0
Pillow==10.1.0

# Database
SQLAlchemy==2.0.23
alembic==1.12.1

# API Client
requests==2.31.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Packaging
pyinstaller==6.2.0

# Optional: для работы с Excel (если нужно)
openpyxl==3.1.2

